requirements.txt
django
Mysqlclient
PyYAML==3.11
